Add the UsaBanner to web components and make the example more interesting
I wanted to be able to show some of the capabilities for the web components in the `cdn-web-components` example
