--- packages/design-system/src/components/web-components/ds-button.ts
@@ -1,20 +1,0 @@
-import { define } from 'preactement';
-import Button from '../Button/Button';
-
-const attributes = [
-  'class-name',
-  'disabled',
-  'href',
-  'is-alternate',
-  'on-dark',
-  'size',
-  'type',
-  'variation',
-  'analytics',
-  'analytics-label-override',
-  'analytics-event-type-override',
-  'analytics-parent-heading',
-  'analytics-parent-type',
-];
-
-define('ds-button', () => Button, { attributes });
