+++ packages/design-system/src/components/web-components/ds-usa-banner.ts
@@ -1,0 +1,6 @@
+import { define } from 'preactement';
+import UsaBanner from '../UsaBanner/UsaBanner';
+
+const attributes = ['class-name', 'id'];
+
+define('ds-usa-banner', () => UsaBanner, { attributes });
