+++ packages/design-system/src/components/web-components/index.ts
@@ -15,3 +15,4 @@ export * from '../i18n';
 export * from './ds-alert';
 export * from './ds-badge';
 export * from './ds-button';
+export * from './ds-usa-banner';
